# lib/api-zod/src/generated/types/bulkInvoiceActionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Collections Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { BulkInvoiceActionInputAction } from './bulkInvoiceActionInputAction';

export interface BulkInvoiceActionInput {
  /** @minItems 1 */
  invoice_ids: string[];
  action: BulkInvoiceActionInputAction;
}