# lib/api-zod/src/generated/types/bulkInvoiceActionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Collections Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { BulkInvoiceActionResponseAction } from './bulkInvoiceActionResponseAction';
import type { Invoice } from './invoice';

export interface BulkInvoiceActionResponse {
  action: BulkInvoiceActionResponseAction;
  requested_count: number;
  updated_count: number;
  skipped_count: number;
  invoices: Invoice[];
}