# lib/api-zod/src/generated/types/collectionTemplatesInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Collections Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { Template } from './template';

export interface CollectionTemplatesInput {
  gentle?: Template;
  firm?: Template;
  serious?: Template;
  final?: Template;
}