# lib/api-zod/src/generated/types/escalationThresholdsInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Collections Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface EscalationThresholdsInput {
  /** @minimum 0 */
  gentle?: number;
  /** @minimum 0 */
  firm?: number;
  /** @minimum 0 */
  serious?: number;
  /** @minimum 0 */
  final?: number;
}