# lib/api-zod/src/generated/types/messageHistory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Collections Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageHistoryEntry } from './messageHistoryEntry';

export type MessageHistory = MessageHistoryEntry[];