# lib/api-zod/src/generated/types/messageHistoryEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Collections Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageHistoryEntryAction } from './messageHistoryEntryAction';
import type { MessageHistoryEntrySource } from './messageHistoryEntrySource';
import type { MessageHistoryEntryTone } from './messageHistoryEntryTone';

export interface MessageHistoryEntry {
  id: string;
  timestamp: Date;
  invoice_id: string;
  customer_name: string;
  tone: MessageHistoryEntryTone;
  source: MessageHistoryEntrySource;
  action: MessageHistoryEntryAction;
  subject: string;
  message: string;
}